requirements.txt
numpy
scipy
matplotlib
ffmpeg
pytorch
typing_extensions